fix: problem with gradle version resolution (#141)
## Description

Refactor how the Gradle dependencies are resolved because as the full
tree is available in both component and stack analysis we can use the
right trees instead of just the api / implementation listings. This
helps resolving the correct versions.

With that it is also possible to include the correct scope (required /
optional) to the sbom



**Related issue (if any):** 

- fixes: https://github.com/trustification/exhort-java-api/issues/138
- fixes: https://github.com/trustification/exhort-java-api/issues/139
- fixes: https://github.com/trustification/exhort-java-api/issues/140

Signed-off-by: Chao Wang <chaowan@redhat.com>